Project structure
- Move handlers to single package
- Create basic structure
