+++ common/logging/logging.go
@@ -1,0 +1,3 @@
+package logging
+
+import "github.com/sirupsen/logrus"
