+++ resources/config.ini
@@ -1,0 +1,3 @@
+[main]
+logfile = /var/log/minimblog.log
+port = 8090
